+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to AOC-Mod will be documented in this file.
 
 ## [Unreleased]
 
+## [0.2.0] - 2025-10-12
+
+### Added
+
+- Added documentation using Sphinx and deployed via readthedocs.
+- Added support for custom, user-generated file templates (more info on [aoc_mod readthedocs](https://readthedocs.org/projects/aoc-mod/aoc_mod_usage.html)).
+- Added ability to run the command-line utility (and all other library utilities) without an authenticated session-id. Without a session id, a user is unable to retrieve puzzle input or submit answers but can receive the first part of the Advent of Code instructions and setup their project challenge.
+- Created a general AocModError exception class to make it easier for library users to detect errors.
+
+### Changed
+
+- Reorganized the project structure to make a bit more sense.
+- Reworked the build system to use Poetry instead of SetupTools.
+- Updated the Python template to be functioning source code rather than a string for better testing.
+- Update testing to reflect new library changes.
+- Replaced all instances of 'os.path' with 'pathlib.Path' objects for better path management
+
+### Documentation
+
+- Updated the README and created other documentation with Sphinx
+- Added or updated function/class docstrings and other comments throughout the codebase
+
 ## [0.1.5] - 2024-11-30
 
 ### Added